+++ scripts/quick_check_imagedl.py
@@ -16,7 +16,7 @@ from imagedl.modules import ImageClientBuilder, BaseImageClient, ImageInfo
 
 
 '''settings'''
-QUERiES = ["Pikachu", "JK", "Cute Animals", "Mountains", "Girls", "Red Panda"]
+QUERiES = ["Pikachu", "JK", "Cute Animals", "Mountains", "Girls", "Red Panda", "James Webb"]
 MAX_SEARCH = 10
 MAX_DL_PER_CLIENT = 10
 RESULTS_ROOT = Path("daily_test_results")
@@ -108,6 +108,8 @@ def main():
                 image_infos = client.search(QUERiES[3], search_limits=MAX_SEARCH, num_threadings=2)
             elif client_name in {'INaturalistImageClient'}:
                 image_infos = client.search(QUERiES[5], search_limits=MAX_SEARCH, num_threadings=2)
+            elif client_name in {'NASAImageClient'}:
+                image_infos = client.search(QUERiES[6], search_limits=MAX_SEARCH, num_threadings=2)
             else:
                 image_infos = client.search(QUERiES[0], search_limits=MAX_SEARCH, num_threadings=2)
             for image_info in image_infos: image_info.raw_data = {}